Cleanup
Updated Email field in PTGOTW writeup to be a select field
Updated Completed field in PTGOTW writeup to be a checkbox field
Restructured files, there's now a tasks and pages folder. Pages is where you define the notion page getting pulled in and write the get method for it - tasks is the logic for interpreting data from notion.